Fix issue with GraphQL, refactor client provider code
